sr(api): move TrainingTriggerTube model in devcomp
