+++ api/lib/domain/models/index.js
@@ -124,7 +124,7 @@ import { TargetProfileSummaryForAdmin } from './TargetProfileSummaryForAdmin.js'
 import { Thematic } from './Thematic.js';
 import { Training } from './Training.js';
 import { TrainingTrigger } from '../../../src/devcomp/domain/models/TrainingTrigger.js';
-import { TrainingTriggerTube } from './TrainingTriggerTube.js';
+import { TrainingTriggerTube } from '../../../src/devcomp/domain/models/TrainingTriggerTube.js';
 import { Tube } from './Tube.js';
 import { Tutorial } from './Tutorial.js';
 import { TutorialEvaluation } from './TutorialEvaluation.js';
