+++ api/tests/tooling/domain-builder/factory/build-training-trigger-for-admin.js
@@ -1,6 +1,6 @@
 import { buildTube } from './build-tube.js';
 import { TrainingTriggerForAdmin } from '../../../../src/devcomp/domain/read-models/TrainingTriggerForAdmin.js';
-import { TrainingTriggerTube } from '../../../../lib/domain/models/TrainingTriggerTube.js';
+import { TrainingTriggerTube } from '../../../../src/devcomp/domain/models/TrainingTriggerTube.js';
 
 const buildTrainingTriggerForAdmin = function ({
   id = 1000,
