+++ api/tests/tooling/domain-builder/factory/build-training-trigger-tube.js
@@ -1,5 +1,5 @@
 import { buildTube } from './build-tube.js';
-import { TrainingTriggerTube } from '../../../../lib/domain/models/TrainingTriggerTube.js';
+import { TrainingTriggerTube } from '../../../../src/devcomp/domain/models/TrainingTriggerTube.js';
 
 const buildTrainingTriggerTube = function ({ id = 1000, tube = buildTube(), level = 8 } = {}) {
   return new TrainingTriggerTube({
