+++ api/tests/tooling/domain-builder/factory/build-training-trigger.js
@@ -1,6 +1,6 @@
 import { buildTube } from './build-tube.js';
 import { TrainingTrigger } from '../../../../src/devcomp/domain/models/TrainingTrigger.js';
-import { TrainingTriggerTube } from '../../../../lib/domain/models/TrainingTriggerTube.js';
+import { TrainingTriggerTube } from '../../../../src/devcomp/domain/models/TrainingTriggerTube.js';
 
 const buildTrainingTrigger = function ({
   id = 1000,
